[CORE][CONFIG] - Created new flood protection for Potion. - Removed flood protector sendMail. - Now FloodProtectorInterval can be a float value. - Fixed bug with augmentetion. - Format RequestRefine.java - Format FloodProtectors.java - Fixed bug with Geodata if target goes behind wall.

+++ gameserver/head-src/com/l2jfrozen/gameserver/util/FloodProtectorConfig.java
@@ -26,7 +26,7 @@ public final class FloodProtectorConfig
 	/**
 	 * Flood protection interval in game ticks.
 	 */
-	public int FLOOD_PROTECTION_INTERVAL;
+	public float FLOOD_PROTECTION_INTERVAL;
 	/**
 	 * Whether flooding should be logged.
 	 */
